fix: lex comments instead of two regexes
Eleven readers in this tree, eleven comment strippers. Nine of them were the same
two regexes over the whole file — remove `/* … */`, remove `//…` — and neither
regex can tell a comment from a string, so the ORDER of the two decides which
half of the tree goes invisible. Both orders are wrong:

  · block-first lets a line comment ending in `/*` — `@girs/*`, `packages/*`,
    `src/*` — pair with the next `*/` below, usually the following JSDoc.
  · line-first lets a block comment CONTAINING a `//` lose its terminator and
    swallow the declaration under it. `npm-registry/src/types.ts` documents a
    field as keyed by a `//host/path/:` prefix; under that order its `authTokens`
    line disappears. The `[^:]` guard does not help — the `//` there follows a
    backtick, not a colon.
  · neither sees a glob pattern in a string. `globSync('**/*.ts')` opens a fake
    block comment in `packages/node/fs/src/glob.spec.ts` whichever way round the
    passes go; the two orders simply blank different halves of that file.

Measured against a stateful scanner as ground truth, over the 3642 tracked JS/TS
sources: block-first hides 7780 code lines in 226 files, line-first hides 3503 in
104. Swapping the order halves a class it cannot close, and leaves the half that
is left looking fixed.

THE CLOSED VERSION ALREADY EXISTED HERE. `check-adapter-import-direction.mjs`
rewrote its stripper as a stateful scanner after a regex read `const re = /[/*]/;`
as an open block comment running to EOF and reported a live violation as exit 0.
That scanner is now `manifest-conformance/lib/strip-comments.mjs`: it knows
strings, template holes and regex literals, keeps every non-comment byte and every
line boundary, and by the same measurement hides 0 code lines and keeps 0 comment
lines. Ten readers share it — the eight whole-file copies, `suite-registration`
(a character scanner that did not lex regex literals, so a `/'/` left string state
open) and `source-graph` (substring-scoped).

Its self-test runs on IMPORT, so every consumer inherits the proof rather than
trusting that someone ran it. A/B, actually run: with the module's `stripComments`
swapped for block-first the import throws naming the `@girs/*` vector; with
line-first it throws naming the `//host/path/:` one. Both vectors carry a trailing
`/** … */` — without a later `*/` the lazy block regex finds no match and neither
bug reproduces.

Two behaviours change where a check was deliberately narrow. `check-adwaita-icon-masks`
and the two anchored readers only stripped WHOLE-LINE `//`, because mid-line is a
URL more often than a comment; the scanner knows which, so a trailing comment now
goes and a `https://` stays. Output is unaffected: all 65 `scripts/check-*.mjs`
keep their exit codes, and every one keeps its output byte for byte.

Numbers in three headers corrected to what re-running them prints: repo-wide
7780/226 of 3642, not 8243/307 of 3641 (that figure counted the comment lines
inside a swallowed span as code); `check-cli-own-version-read`'s own corpus
2448/47, not 2446/46.
