+++ scripts/check-cli-own-version-read.mjs
@@ -64,6 +64,7 @@ import {
     TS_SOURCE_EXTENSIONS,
     sourceExtensionRe,
 } from '../packages/infra/manifest-conformance/lib/source-extensions.mjs';
+import { stripComments } from '../packages/infra/manifest-conformance/lib/strip-comments.mjs';
 
 const REPO_ROOT = join(dirname(fileURLToPath(import.meta.url)), '..');
 const CLI_SRC = join(REPO_ROOT, 'packages/infra/cli/src');
@@ -128,22 +129,18 @@ if (files.length === 0) {
     process.exit(1);
 }
 
-// Blank out comment bodies while preserving line numbering: the rule this check
-// enforces is precisely the kind a comment needs to QUOTE in order to explain
-// itself, and a guard that fires on its own rationale trains people to delete
-// the rationale.
+// Comments are removed, line numbering intact: the rule this check enforces is precisely
+// the kind a comment needs to QUOTE in order to explain itself, and a guard that fires on
+// its own rationale trains people to delete the rationale.
 //
-// LINE COMMENTS FIRST, and the order is load-bearing. A line comment ending in `/*` —
-// `@gjsify/*`, `packages/*`, `src/*`, all common here — otherwise pairs with the next
-// `*/` anywhere below, usually the opening of the following JSDoc, and everything
-// between is blanked. Measured over this check's own corpus before the swap: 2446 code
-// lines across 46 of its 285 files were invisible to it. A guard that cannot see two
-// thousand lines of what it guards is not a weak guard, it is a decoration.
-function stripComments(text) {
-    return text
-        .replace(/(^|[^:])\/\/[^\n]*/g, (m, lead) => lead + ' '.repeat(m.length - lead.length))
-        .replace(/\/\*[\s\S]*?\*\//g, (m) => m.replace(/[^\n]/g, ' '));
-}
+// The reader is `scripts/strip-comments.mjs`, a lexical scanner, because no PAIR of
+// regexes can do this. Removing block comments first let a line comment ending in `/*` —
+// `@gjsify/*`, `packages/*`, `src/*`, all common here — pair with the next `*/` below and
+// blank everything between; removing line comments first let a block comment containing a
+// `//` lose its terminator and swallow the declaration under it instead. Measured over
+// this check's own corpus, block-first hid 2448 code lines across 47 of its 285 files. A
+// guard that cannot see two thousand lines of what it guards is not a weak guard, it is a
+// decoration.
 
 const offenders = [];
 for (const file of files) {
